remove spurious type conversions
If my understanding of modular arithmetic and wrapping behavior is correct, conversion to u128 is both unnecessary, and slower

+++ src/lib.rs
@@ -235,7 +235,7 @@ impl<T: ?Sized> Bloom<T> {
             hashes[k_i as usize] = hash;
             hash
         } else {
-            (hashes[0] as u128).wrapping_add((k_i as u128).wrapping_mul(hashes[1] as u128)) as u64
+            (hashes[0]).wrapping_add((k_i as u64).wrapping_mul(hashes[1]))
                 % 0xFFFF_FFFF_FFFF_FFC5u64 //largest u64 prime
         }
     }
